User-only components SSR only + path header fix
Components only avaialble to certain users are now rendered client-side only without hydration errors. This is because the cookie is not available to the server making the request. We may wish to figure out a way to change this, but I'm not sure we have many options.
Path header is now the original route URL, not just the same url as is being requested. Bug fix.